Split the language model register tab panel into the separated js file.
